refactor(type_checker): extract BinaryOp into binop.rs
